pyproject.toml
[tool.poetry]
name = "djangoapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
Django = "3.2.6"
djangorestframework = "3.12.4"
gensim = "4.0.1"
keras = "2.6.0"
django-phonenumber-field = "^5.2.0"
phonenumbers = "^8.12.31"
scikit-learn = "0.24.2"
tensorflow = "2.6.0"
scipy = "^1.7.1"
nltk = "^3.6.2"
sentence-transformers = "2.0.0"
pymorphy2 = "0.9.1"
python-dotenv = "^0.19.0"
psycopg2 = "^2.9.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
